// packages/starlight-asyncapi/libs/schemas/schema.ts
import { z } from 'astro/zod'

import { getOperationSidebarGroups } from '../navigation'
import type { NormalizedAsyncAPIDocument } from '../normalize'
import type { TranslateFn } from '../operation'
import { getSchemaBaseLink, stripLeadingAndTrailingSlashes } from '../path'
import { makeSidebarGroup, makeSidebarLink, type SidebarGroup, type StarlightAsyncAPISidebarGroup } from '../starlight'
import type { StarlightAsyncAPIContext } from '../vite'

export const SchemaConfigSchema = z.object({
  /**
   * The AsyncAPI route base path containing the generated documentation.
   * @example 'events/orders'
   */
  base: z.string().min(1).transform(stripLeadingAndTrailingSlashes),
  /**
   * The AsyncAPI schema path or URL.
   */
  schema: z.string().min(1),
  /**
   * The generated sidebar group configuration.
   */
  sidebar: z
    .object({
      /**
       * Wheter the generated documentation sidebar group should be collapsed by default.
       * @default true
       */
      collapsed: z.boolean().default(true),
      /**
       * The generated documentation sidebar group label.
       * @defaultValue
       * Defaults to the AsyncAPI document title.
       */
      label: z.string().optional(),
      /**
       * The optional sidebar group generated by the `createAsyncAPISidebarGroup()` function that will contain the
       * generated documentation pages.
       */
      group: (z.any() as z.ZodType<StarlightAsyncAPISidebarGroup>).optional(),
      /**
       * The generated documentation operations sidebar links configuration.
       */
      operations: z
        .object({
          /**
           * Defines if the sidebar should display badges next to operation links with the associated action
           * (`send`/`receive`) and protocol.
           * @default false
           */
          badges: z.boolean().default(false),
          /**
           * Whether the operation sidebar link labels should use the operation ID, channel address, or summary.
           * @default 'summary'
           */
          labels: z.enum(['operationId', 'channel', 'summary']).default('summary'),
          /**
           * Defines the sorting method for the operation sidebar links.
           * @default 'document'
           */
          sort: z.enum(['alphabetical', 'document']).default('document'),
        })
        .prefault({}),
      /**
       * The generated documentation tags sidebar groups configuration.
       */
      tags: z
        .object({
          /**
           * Defines the sorting method for the tag sidebar groups.
           * @default 'document'
           */
          sort: z.enum(['alphabetical', 'document']).default('document'),
        })
        .prefault({}),
    })
    .prefault({}),
})

export function getSchemaSidebarGroups(
  pathname: string,
  schema: Schema,
  context: StarlightAsyncAPIContext,
  t: TranslateFn,
  defaultGroupLabel: string,
): [label: string, group: SidebarGroup] {
  const { config, document } = schema

  return [
    config.sidebar.group?.label ?? defaultGroupLabel,
    makeSidebarGroup(
      config.sidebar.label ?? document.info.title,
      [
        makeSidebarLink(pathname, t('nav.overview'), getSchemaBaseLink(config, context)),
        ...getOperationSidebarGroups(pathname, schema, context, t),
      ],
      config.sidebar.collapsed,
    ),
  ]
}

export type StarlightAsyncAPISchemaConfig = z.output<typeof SchemaConfigSchema>

export interface Schema {
  config: StarlightAsyncAPISchemaConfig
  document: NormalizedAsyncAPIDocument
}